+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/tobe/HandlerMappings.java
@@ -1,8 +1,6 @@
-package com.techcourse.support.web.mapping;
+package webmvc.org.springframework.web.servlet.mvc.tobe;
 
 import jakarta.servlet.http.HttpServletRequest;
-import webmvc.org.springframework.web.servlet.mvc.tobe.AnnotationHandlerMapping;
-import webmvc.org.springframework.web.servlet.mvc.tobe.HandlerMapping;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -16,14 +14,9 @@ public class HandlerMappings {
         this.mappings = new ArrayList<>();
     }
 
-    public void init() {
-        final AnnotationHandlerMapping annotationHandlerMapping = new AnnotationHandlerMapping();
-        annotationHandlerMapping.initialize();
-        final ManualHandlerMapping manualHandlerMapping = new ManualHandlerMapping();
-        manualHandlerMapping.initialize();
-
-        mappings.add(annotationHandlerMapping);
-        mappings.add(manualHandlerMapping);
+    public void init(final HandlerMapping... mappings) {
+        this.mappings.add(new AnnotationHandlerMapping());
+        this.mappings.addAll(List.of(mappings));
     }
 
     public Optional<Object> getHandler(final HttpServletRequest request) {
